# lib/api-zod/src/generated/types/crawlJobStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CrawlJobStatus {
  jobId: number;
  status: string;
  /** @nullable */
  extractorVersion?: string | null;
  pagesFound: number;
  pagesCrawled: number;
  pagesFailed: number;
  pagesSkipped: number;
  chunksExtracted: number;
  /** @nullable */
  errorSummary?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}